// src/excercises/enums/excercise-errors.enum.ts
export enum ExcerciseError {
  ALTERNATIVE_LENGTH = 'Alternativa errónea',
  EXCERCISE_STATEMENT = 'Enunciado del ejercicio inválido',
  EXCERCISE_TYPE = 'Tipo de ejercicio incorrecto',
  SELECTION_ANSWER = 'Respuesta no escogida',
  FILL_ANSWER = 'Respuesta no escrita',
  DRAG_GROUPS = 'Grupos no definidos',
  TEACHER = 'Profesor no escogido',
  COURSE = 'Curso incorrecto',
  PHOTO = 'Foto incorrecta',
}